Added the network handling code
The prototype now tries to connect to the cairo-display server (which
is not written yet) and receives a cairo script from it.
For prototyping purposes, the protocol is similar to heredoc: the lines
are read until an "EOF" marker is received, at which point the client
displays the result on the screen and sends "OK" back to the server.