Use copy-page instruction to display data
Instead of relying on the underlying transport protocol, it is better to use
the copy-page instruction of the cairo script. So the implementation for the
HereDoc-style protocol is not needed anymore and should be removed in the
future.

Upon receiving a copy-page instruction, the cairo script interpreter will call
our hook function, which will create the Android bitmap and pass it to the
Java listener. This greatly simplifies the Java part of the code.