Fix multiple creation of the CairoClient
When switching the display orientation, the Activity.onCreate() is
called multiple times with a corresponding calls to the
Activity.onDestroy(), but this was not accounted for and eventually
lead to multiple connections being made to the server. The error
became apparent only ofter the display became the server.
The CairoClient is now owned by the DisplayActivity instead of the
CairoDisplay, and is being managed by it.